refactor(desktop): extract the native application menu from main.mjs
Menu template installation, the native-menu → renderer forwarding actions
(settings, check-updates, sidebar toggle, zoom via the renderer's font-zoom
pathway), and Windows/Linux menu-bar visibility move to
electron/app-menu.mjs as a createApplicationMenu factory with injected
appName/docsUrl/getWindow. The NATIVE_MENU_* channel names are duplicated
constants in preload.mjs already (no new coupling); menu visibility state
is now module-private.

Tested:
- node --check both; typecheck:electron clean; check:electron 50 methods;
  desktop node:test 18 pass
- LIVE: booted the Electron dev app — menu installs at startup,
  __setApplicationMenuVisible roundtrips true/false through the factory,
  bridge still serves (9 workspaces).
